re-add framer-motion, move colormodetoggle into components folder
