+++ package.json
@@ -15,6 +15,7 @@
     "@types/react-datepicker": "4.3.4",
     "@types/react-dom": "17.0.20",
     "date-fns": "2.30.0",
+    "framer-motion": "4.1.17",
     "history": "5.2.0",
     "react": "17.0.2",
     "react-datepicker": "4.6.0",
